refactor: 3단계 리팩토링
- Strategy 인터페이스 추상화
- Controller 에러 처리 깊이 변경
- 패키지 정리
- 메소드 정리
