+++ src/main/java/lotto/strategy/AutomaticStrategy.java
@@ -1,0 +1,10 @@
+package lotto.strategy;
+
+import java.util.List;
+
+public interface AutomaticStrategy {
+
+    int SIZE = 6;
+
+    List<Integer> choose(List<Integer> numbers);
+}
