+++ src/main/java/lotto/strategy/SublistStrategy.java
@@ -1,10 +1,10 @@
-package lotto;
+package lotto.strategy;
 
 import java.util.ArrayList;
 import java.util.List;
 
 public class SublistStrategy implements AutomaticStrategy {
-    public ArrayList<Integer> choose(List<Integer> numbers) {
+    public List<Integer> choose(List<Integer> numbers) {
         return new ArrayList<>(numbers.subList(0, SIZE));
     }
 }
