+++ src/test/java/step2/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package step2;
 
 import static org.assertj.core.api.Assertions.*;
